Reorganize structure to abstract the kernel out
Much less code in the simple driver now, and have fairly easily-defined
structure for the kernels to follow.
Next step in this process would be to make a virtual KernelCPUContext
class to derive the actual kernel objects from.
Is that too far already? I don't think so, as it will reduce code
repetition and possible room for coding mistakes.