Fix issue with GPU usage on other arch's
The first sign of a problem was that the code would fail at the
GetFuncAttributes when getting register counts.
The cudaPrintLastError() said that the device function was invalid.
Research (aka StackOverflow) said the problem was not compiling
for the correct compute capability, which makes sense, as the other
architectures I am testing on have different compute capabilities.
See https://stackoverflow.com/a/62931012 for the answer I went with.