Fix bug where the uninit() doesn't empty indices
Just a remnant from when I was building for a very simple kernel.
This current version is much closer to as general as possible.
Only (big) thing I am missing (that I can think of) is having the
indices and value generation be more generic, as in come from a
set of predefined patterns. As that would be a bunch of either
pre-constructed arrays or function pointers to generate the values,
these could easily be passed as references at run time and not
have to be compiled.