Basic changes to plot something
Hardcoded values; plan to back out and generalize from here.
Need to really have bandwidth be the cut off, probably finding
maximum bandwidth (or minimum time) for a given computation and then
plotting the values of each of the different run configurations that
have the lowest occupancy that surpasses ~90-95% of that maximum
bandwidth.
Essentially, trying to boil down the occupancy need reduction into a
single metric.
First, will really need to add throughput to the output files.