Change return type `get_total_bytes_processed()`
Just using `int` was not large enough for A100, as had large value of N,
and the many loops caused a high enough value of bytes to overflow
into negative numbers, which is obviously incorrect.
Since it is only used in one spot (per driver instance), the extra
memory of the `long long` should not matter at all.