Add individual funcs for plotting each kernel type
Started with computational intensity, since it seemed like the most
complex one so far.
This seems to have worked failry well.
I plan to replicate this basic structure for all kernels to come.
Not as generic as could have, but better than overengineering.

Also includes a `kernel_type_names` dictionary, which seemed necessary.
I plan to abstract out part of the plotting stuff.