Test block sizes 256, 512, and 1024
To see impact of current indexing scheme.
The large the block, the more the striding, but
it is still limited to only a block's worth of data.
Need to implement some kind of randomization via the index generation
stuff, but that adds more reads to the logic.
Maybe do some simpler striding? I still need to keep it within some
control...