Move timing logic to derived kernel side
Noticed when debugging that the timing was counting the whole
`local_execute()` execution, not just the kerenl launch.
This makes sense, but I do *not* want to have the local logic impact
the timing, mainly due to the recent attempts to increase dynamic shared
memory allotment.
Putting the timing tightly wrapped around the kernel invocation should
provide the smallest granularity in timing possible.