Fix memory build up bug
I misunderstood what vector::clear() does.
It does **not** reduce the actual capacity (memory allocation) of the
container, just remove the values (I guess make them undefined?).
Change this functionality to swap with an empty vector, which should
(hopefully) reset the capacity to 0.
This was recommended by several people on SO
https://stackoverflow.com/questions/1111078/reduce-the-capacity-of-an-stl-vector
and a reference to Stroustrup's recommendation is made, as well.
Not that the last bit really matters, but I thought it was interesting.