Update kernel for more general ILP/elements usage
This is more complete in that it allows for ILP > elements.
Performance is currently pretty good on local desktop, though I see the
register usage is rather high for large number of elements.
I don't really know why it should be that way, but the compiler usually
does pretty good.