Add ability to output extra cfg info from kernels
Instead of having the info need to be extracted from the filename in
postprocessing, put the info directly into the file contents.
This is a better alternative (actually based off work I did in another
repo for another project).
I am positive there are better ways of doing this, but this method
works, and I am in no mood to redesign things yet.