Fix `uncoalesced_access_shuffle_size()`
The previous problem was that the indices generated were not correct.
My checking code will appear soon.
This essentially provides a logical loop structure to compute the
indirection indices.
A shuffle block is broken into multiple "scans" composed of multiple
warps, and this structure seems to make sense.
However, this is probably not useful for shuffle sizes < 1024.
This may also be somewhat difficult to put into the older uncoalesced
reuse kernel's index computation structure; though I have an idea that
it shouldn't be too bad.