Reduce clutter in `check_contraction()`
There honestly isn't much of a test I can really do for this...
Not that I am aware of, at least.
I guess I can loop over it in a warp-centric way now, but I am unsure
if this would just be a reinterpretation of the math I already did?