Current draft of SpMV draft
I have neglected to commit for a couple days, so wanted to at least put
the current version here.
Since this didn't fit the rest of the framework
(`MicrobenchmarkDriver` and parent
`KernelContext` classes), I had to make a standalone version.
I will probably need to make a second version for more accurate usage.