Remove dependence on SM count for min division
The old usage was meant to ensure there was exactly equal work to do on all SMs
in total, even though we can't control which SM gets which block (directly, anyways). 
This new change may incur some additional variance in timing, but I honestly doubt it. 
Hopefully, this will run first try, but we will see...