Add `tests` subfolder to project
This is just a proof of concept to add googletest to the project.
Long-term, I would prefer to have my tests in a `tests` subdirectory
of their place in the project.
But for now, I have so few tests in mind that I think they could
almost all fit in this global `tests` directory.