Remove build of spmv on windows
I have some terrible issue getting the `force_global_load()` function to
work correctly.
Since the work I am trying to do on Windows is just restructuring the
"business logic" of the benchmark driver, and I
plan to use that base code in the spmv anyways, I don't feel that I
need to update the spmv during this work.