Update DirectoryFinderTests for parent finding
Actually was updating this and the actual code at the same time.
I'm trying to do TDD, but there is still pretty close coupling between
the test development and the actual features here.
Mostly because I am not as experienced with this filesystem stuff as
some other stuff in this codebase (maybe an excuse...)