Add nvbench to CMakeLists file; not an interface
This is problaby not the way to go, but if using a built library
(nvbench), will likely need this to be a built library as well?
Also, still part of this project will need to be header for the
templates?
I still don't know what I'm doing, but I'm trying....